+++ src/value.rs
@@ -3,12 +3,12 @@ use std::mem;
 use std::slice;
 
 extern {
-  fn fs_alloc(ptr_count: u16, aux_count: u16) -> *mut c_void;
+  fn fs_alloc(ptr_count: usize, aux_count: usize) -> *mut c_void;
   fn fs_dealloc(lay: *mut c_void);
-  fn fs_ptr_count(lay: *mut c_void) -> u16;
-  fn fs_aux_count(lay: *mut c_void) -> u16;
-  fn fs_get_ptr(lay: *mut c_void, offset: u16) -> *mut c_void;
-  fn fs_set_ptr(lay: *mut c_void, offset: u16, ptr: *mut c_void) -> *mut c_void;
+  fn fs_ptr_count(lay: *mut c_void) -> usize;
+  fn fs_aux_count(lay: *mut c_void) -> usize;
+  fn fs_get_ptr(lay: *mut c_void, offset: usize) -> *mut c_void;
+  fn fs_set_ptr(lay: *mut c_void, offset: usize, ptr: *mut c_void) -> *mut c_void;
   fn fs_aux(lay: *mut c_void) -> *mut u8;
 }
 
@@ -18,7 +18,7 @@ pub struct Ref<'a> {
 }
 
 impl<'a> Ref<'a> {
-  pub fn get_ptr(&self, offset: u16) -> Option<Ref<'a>> {
+  pub fn get_ptr(&self, offset: usize) -> Option<Ref<'a>> {
     if offset >= self.ptr_count() {
       None
     } else {
@@ -32,7 +32,7 @@ impl<'a> Ref<'a> {
     }
   }
 
-  pub fn set_ptr(&self, offset: u16, ptr: &Ref<'a>) {
+  pub fn set_ptr(&self, offset: usize, ptr: &Ref<'a>) {
     // FIXME: Check that GCs are the same.
     if offset < self.ptr_count() {
       unsafe { fs_set_ptr(self.lay, offset, ptr.lay) };
@@ -41,7 +41,7 @@ impl<'a> Ref<'a> {
 
   pub fn aux(&self) -> &mut [u8] {
     unsafe {
-      slice::from_raw_parts_mut(fs_aux(self.lay), self.aux_count() as usize)
+      slice::from_raw_parts_mut(fs_aux(self.lay), self.aux_count())
     }
   }
 
@@ -55,11 +55,11 @@ impl<'a> Ref<'a> {
     }
   }
 
-  pub fn ptr_count(&self) -> u16 {
+  pub fn ptr_count(&self) -> usize {
     unsafe { fs_ptr_count(self.lay) }
   }
 
-  pub fn aux_count(&self) -> u16 {
+  pub fn aux_count(&self) -> usize {
     unsafe { fs_aux_count(self.lay) }
   }
 }
@@ -85,7 +85,7 @@ impl GC {
     GC{}
   }
 
-  pub fn alloc<'a>(&'a self, ptr_count: u16, aux_count: u16) -> Ref<'a> {
+  pub fn alloc<'a>(&'a self, ptr_count: usize, aux_count: usize) -> Ref<'a> {
     unsafe {
       // FIXME: Tell GC to retain root.
       let lay = fs_alloc(ptr_count, aux_count);
